+++ packages/plugin-lava-player/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @discordx/plugin-lava-player
 
+## 2.5.5
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @discordx/lava-queue@4.0.6
+
 ## 2.5.4
 
 ### Patch Changes