+++ packages/plugin-ytdl-player/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @discordx/plugin-ytdl-player
 
+## 2.3.5
+
+### Patch Changes
+
+- update queue
+
+## 2.3.4
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @discordx/music@6.3.1
+
 ## 2.3.3
 
 ### Patch Changes